Move helper code into test fixture
