+++ base/BUILD
@@ -30,6 +30,8 @@ cc_library(
 cc_test(
   name = "test_main",
   srcs = [
+    "test/helper.h",
+    "test/helper.cc",
     "test/main.cc",
     "test/task_posting.cc",
     "test/bind_task_loop_to_current_thread.cc"
